Moved AfterData function to viewmodel
